server/requirement.txt
fastapi
uvicorn[standard]
python-multipart

langchain
langchain-community
langchain-core

pinecone

langchain-google-genai

PyPDF

pydantic
requests
tqdm

loguru

python-dotenv


